+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/testutils/number/commonByteTests.kt
@@ -1,40 +1,6 @@
 package xyz.lbres.kotlinutils.testutils.number
 
-import xyz.lbres.kotlinutils.internal.constants.Suppressions
-import xyz.lbres.kotlinutils.testutils.runWithFailMessage
 import kotlin.test.assertEquals
-
-val byteNotZeroValues: List<Pair<List<Byte>, List<Byte>>> = listOf(
-    emptyList<Byte>() to emptyList(),
-    listOf(0, 0, 0).map(Int::toByte) to emptyList(),
-    listOf(1, 2, 0, 4, 0, 0, 5).map(Int::toByte) to listOf(1, 2, 4, 5),
-    listOf(-1, 1, 0).map(Int::toByte) to listOf(-1, 1),
-    listOf(1, 4, 100, 19, 5).map(Int::toByte) to listOf(1, 4, 100, 19, 5)
-)
-val intNotZeroValues: List<Pair<List<Int>, List<Int>>> = listOf(
-    emptyList<Int>() to emptyList(),
-    listOf(0, 0, 0) to emptyList(),
-    listOf(1, 2, 0, 4, 0, 0, 5) to listOf(1, 2, 4, 5),
-    listOf(-1, 1, 0) to listOf(-1, 1),
-    listOf(1, 4, 1000, 19, 5) to listOf(1, 4, 1000, 19, 5),
-)
-
-@Suppress(Suppressions.UNCHECKED_CAST)
-inline fun <reified S, T> testFilterNotZeroGeneric(createValues: (List<S>) -> T, crossinline filterNotZero: (T) -> List<S>) {
-    val notZeroValues: List<Pair<List<S>, List<S>>> = when (S::class) {
-        Byte::class -> byteNotZeroValues
-        Int::class -> intNotZeroValues
-        else -> emptyList()
-    } as List<Pair<List<S>, List<S>>>
-
-    for (pair in notZeroValues) {
-        val values: T = createValues(pair.first)
-        val expected: List<S> = pair.second
-        runWithFailMessage("Checking $values with expected result $expected") {
-            assertEquals(expected, filterNotZero(values))
-        }
-    }
-}
 
 fun <T> testSumByte(createValues: (List<Byte>) -> T, sum: (T) -> Byte) {
     var values = createValues(emptyList())
